requirements.txt
flask
requests
yara-python
python-magic
temp